chore: moved bulk of the register form logic to a custom hook. put back the egister:true in the navigation. and fixed the userID typing
