+++ src/components/auth/RegisterForm.tsx
@@ -1,147 +1,17 @@
-import { useState } from "react";
 import { motion, AnimatePresence } from "framer-motion";
-import { useNavigate } from "@tanstack/react-router";
-import { showToast } from "../../utils/toast";
-import { AccountService } from "../../services/account.service";
+import { useRegisterForm } from "../../hooks/useRegisterForm";
 const RegisterForm: React.FC = () => {
-  const navigate = useNavigate();
-  const [currentStep, setCurrentStep] = useState(1);
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-    confirmPassword: "",
-    surname: "",
-    firstname: "",
-    middlename: "",
-    phoneNumber: "",
-    country: "",
-    province: "",
-    city: "",
-    nationalID: "",
-    birthdate: "",
-    nationality: "",
-    account_type: "",
-  });
-  const [fileName, setFileName] = useState("Upload ID");
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const target = e.target as HTMLInputElement;
-    const file = target?.files?.[0];
-    if (file) setFileName(file.name);
-  };
-
-  const nextStep = () => {
-    setCurrentStep((prevStep) => prevStep + 1);
-  };
-  const prevStep = () => {
-    setCurrentStep((prevStep) => prevStep - 1);
-  };
-  const validateForm = () => {
-    if (currentStep === 1) {
-      return formData.surname && formData.firstname && formData.middlename && formData.nationality;
-    } else if (currentStep === 2) {
-      return formData.email && formData.password && formData.confirmPassword;
-    } else if (currentStep === 3) {
-      return formData.phoneNumber && formData.country && formData.province && formData.city;
-    } else if (currentStep === 4) {
-      return formData.nationalID && formData.birthdate && formData.account_type;
-    }
-    return true;
-  };
-  const handleNext = (e: React.MouseEvent) => {
-    e.preventDefault();
-    if (validateForm()) {
-      nextStep();
-    } else {
-      showToast.error("Please fill in all fields");
-    }
-  };
-  const handleEdgeCases = () => {
-    if (!validateForm()) {
-      showToast.error("Please fill in all fields");
-      return;
-    }
-    if (formData.password !== formData.confirmPassword) {
-      showToast.error("Passwords do not match");
-      return;
-    }
-    if (formData.account_type === "") {
-      showToast.error("Please select an account type");
-      return;
-    }
-    if (formData.nationalID === "") {
-      showToast.error("Please upload your ID");
-      return;
-    }
-    if (formData.birthdate === "") {
-      showToast.error("Please select your birthdate");
-      return;
-    }
-    if (formData.phoneNumber === "") {
-      showToast.error("Please enter your phone number");
-      return;
-    }
-    if (formData.country === "") {
-      showToast.error("Please enter your country");
-      return;
-    }
-    if (formData.province === "") {
-      showToast.error("Please enter your province");
-      return;
-    }
-    if (formData.city === "") {
-      showToast.error("Please enter your city");
-      return;
-    }
-    if (formData.surname === "") {
-      showToast.error("Please enter your surname");
-      return;
-    }
-    if (formData.firstname === "") {
-      showToast.error("Please enter your firstname");
-      return;
-    }
-    if (formData.middlename === "") {
-      showToast.error("Please enter your middlename");
-      return;
-    }
-    if (isLoading) return;
-  };
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    handleEdgeCases();
-    setIsLoading(true);
-    console.log(formData);
-    handleRegister();
-  };
-
-  const handleRegister = async () => {
-    const address = `${formData.country}, ${formData.province}, ${formData.city}`;
-    try {
-      await AccountService.register(
-        formData.surname,
-        formData.firstname,
-        formData.middlename,
-        formData.email,
-        formData.password,
-        formData.phoneNumber,
-        address,
-        formData.nationalID,
-        formData.birthdate,
-        formData.nationality,
-        formData.account_type
-      );
-      showToast.success("Registration successful");
-      console.log("Registration OK");
-      navigate({ to: "/welcome/signin" });
-    } catch (error) {
-      showToast.error("Registration failed");
-      console.error("Registration error:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const {
+    currentStep,
+    formData,
+    setFormData,
+    fileName,
+    isLoading,
+    handleFileChange,
+    prevStep,
+    handleNext,
+    handleSubmit,
+  } = useRegisterForm();
 
   function renderStep() {
     switch (currentStep) {
