+++ src/types/contexts.types.ts
@@ -2,7 +2,7 @@ export type AuthContextType = {
   loading: boolean;
   isAuthenticated: boolean;
   username: string | null;
-  userId: number | null;
+  userId: string | null;
   tokenExp: number | null;
   login: (email: string, password: string) => Promise<boolean>;
   logout: () => Promise<boolean>;
